ft: factor out graphql client
 - factor out graphql client

[Delivers #170843732]
