+++ frontend/lib/graphqlClient.ts
@@ -1,0 +1,15 @@
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { ApolloClient } from 'apollo-client';
+import { createHttpLink } from 'apollo-link-http';
+
+const httpLink = createHttpLink({
+    uri: 'http://localhost:8000/graphql/',
+    // uri: 'http://10.20.42.55:8000/graphql/'
+});
+
+const client = new ApolloClient({
+    link: httpLink,
+    cache: new InMemoryCache(),
+});
+
+export default client;
